backend/agents/pyproject.toml
[project]
name = "veritas-ai-agent"
version = "0.1.0"
description = "The Veritas AI Agent evaluates financial reports."
authors = [
    { name = "Orkhan Gasimli", email = "orkhan.gasimli@gmail.com" },
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # User dependencies
    "PyYAML>=6.0.1",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6,<1.0.0",
    "ty>=0.0.1a0",
    "codespell>=2.2.0,<3.0.0",
]

[dependency-groups]
dev = [
    "google-adk[eval]>=1.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
deployment = [
    "absl-py>=2.2.1",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006", "E402"] # ignore line too long, too complex
                                     
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ty]
# ty is Astral's Rust-based type checker (same team as ruff/uv)
# See: https://docs.astral.sh/ty/

[tool.ty.environment]
python-version = "3.10"

[tool.ty.src]
exclude = [".venv/**"]

[tool.ty.rules]
# Ignore common issues with third-party libraries and dynamic code patterns
unresolved-import = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb,**/ifrs_disclosure_checklist.yaml"


[tool.ruff.lint.isort]
# Categorizes the local package for correct import sorting
known-first-party = ["veritas_ai_agent"]

[tool.hatch.build.targets.wheel]
# Ensures Hatchling finds the package folder correctly
packages = ["veritas_ai_agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"